+++ app/src/main/java/com/freetalk/presenter/fragment/chat/ChatFragment.kt
@@ -1,15 +1,15 @@
-package com.freetalk.presenter.fragment
+package com.freetalk.presenter.fragment.chat
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.freetalk.databinding.FragmentChatBinding
 import com.freetalk.databinding.FragmentHomeBinding
-import com.freetalk.databinding.FragmentMyPageBinding
 
-class MyPageFragment: Fragment() {
-    private var _binding: FragmentMyPageBinding? = null
+class ChatFragment: Fragment() {
+    private var _binding: FragmentChatBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
@@ -17,7 +17,7 @@ class MyPageFragment: Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentMyPageBinding.inflate(inflater, container, false)
+        _binding = FragmentChatBinding.inflate(inflater, container, false)
         return binding.root
     }
 }
