+++ app/src/main/java/com/freetalk/presenter/fragment/mypage/MyPageFragment.kt
@@ -1,15 +1,15 @@
-package com.freetalk.presenter.fragment
+package com.freetalk.presenter.fragment.mypage
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import com.freetalk.databinding.FragmentHomeBinding
+import com.freetalk.databinding.FragmentMyPageBinding
 
-class HomeFragment: Fragment() {
-    private var _binding: FragmentHomeBinding? = null
+class MyPageFragment: Fragment() {
+    private var _binding: FragmentMyPageBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
@@ -17,8 +17,7 @@ class HomeFragment: Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentHomeBinding.inflate(inflater, container, false)
-        (activity as AppCompatActivity).setSupportActionBar(binding.appBar)
+        _binding = FragmentMyPageBinding.inflate(inflater, container, false)
         return binding.root
     }
 }
